sparql_service::srdf_data

Struct RdfData

pub struct RdfData { /* private fields */ }

Generic abstraction that represents RDF Data which can be behind SPARQL endpoints or an in-memory graph or both The triples in RdfData are taken as the union of the triples of the endpoints and the in-memory graph

Implementations

impl RdfData

pub fn new() -> RdfData

pub fn reset(&mut self)

pub fn with_rdf_data_config( self, rdf_data_config: &RdfDataConfig, ) -> Result<Self, RdfDataError>

pub fn check_store(&mut self) -> Result<(), RdfDataError>

Checks if the Store has been initialized

By default, the RDF Data Store is not initialized as it is expensive and is only required for SPARQL queries

pub fn from_graph(graph: InMemoryGraph) -> Result<RdfData, RdfDataError>

Creates an RdfData from an in-memory RDF Graph

pub fn clean_all(&mut self)

pub fn graph(&self) -> Option<&InMemoryGraph>

Get the in-memory graph

pub fn graph_prefixmap(&self) -> PrefixMap

pub fn clean_graph(&mut self)

Cleans the in-memory graph

pub fn from_str( data: &str, format: &RDFFormat, base: Option<&str>, reader_mode: &ReaderMode, ) -> Result<RdfData, RdfDataError>

pub fn merge_from_reader<R: Read>( &mut self, read: &mut R, source_name: &str, format: &RDFFormat, base: Option<&str>, reader_mode: &ReaderMode, ) -> Result<(), RdfDataError>

Merge the in-memory graph with the graph read from a reader

pub fn from_endpoint(name: &str, endpoint: SparqlEndpoint) -> RdfData

Creates an RdfData from an endpoint

pub fn add_endpoint(&mut self, name: &str, endpoint: SparqlEndpoint)

Adds a new endpoint to the list of available endpoints

pub fn use_endpoints(&self) -> &HashMap<String, SparqlEndpoint>

pub fn endpoints(&self) -> &HashMap<String, SparqlEndpoint>

pub fn use_endpoint(&mut self, name: &str, endpoint: SparqlEndpoint)

pub fn dont_use_endpoint(&mut self, name: &str)

pub fn endpoints_to_use(&self) -> impl Iterator<Item = (&str, &SparqlEndpoint)>

pub fn prefixmap_in_memory(&self) -> PrefixMap

Gets the PrefixMap from the in-memory graph

pub fn show_blanknode(&self, bn: &OxBlankNode) -> String

pub fn show_literal(&self, lit: &OxLiteral) -> String

pub fn serialize<W: Write>( &self, format: &RDFFormat, writer: &mut W, ) -> Result<(), RdfDataError>

pub fn find_endpoint(&self, name: &str) -> Option<SparqlEndpoint>

pub fn all_triples( &self, ) -> Result<impl Iterator<Item = OxTriple>, RdfDataError>

Gets all the triples from the in-memory graph and the endpoints This operation can be very expensive if the endpoints contain a lot of data